+++ TimberCommons/IrrigationSystem/Configurator.cs
@@ -1,0 +1,27 @@
+﻿// Timberborn Mod: Timberborn Commons
+// Author: igor.zavoychinskiy@gmail.com
+// License: Public Domain
+
+using Bindito.Core;
+using TimberApi.ConfiguratorSystem;
+using TimberApi.SceneSystem;
+using Timberborn.Growing;
+using Timberborn.TemplateSystem;
+
+namespace IgorZ.TimberCommons.IrrigationSystem {
+
+[Configurator(SceneEntrypoint.InGame)]
+// ReSharper disable once UnusedType.Global
+sealed class Configurator : IConfigurator {
+  public void Configure(IContainerDefinition containerDefinition) {
+    containerDefinition.MultiBind<TemplateModule>().ToProvider(ProvideTemplateModule).AsSingleton();
+  }
+
+  static TemplateModule ProvideTemplateModule() {
+    var builder = new TemplateModule.Builder();
+    builder.AddDecorator<Growable, GrowthRateModifier>();
+    return builder.Build();
+  }
+}
+
+}