+++ TimberCommons/WaterService/Configurator.cs
@@ -3,6 +3,7 @@
 // License: Public Domain
 
 using Bindito.Core;
+using IgorZ.TimberDev.Logging;
 using TimberApi.ConfiguratorSystem;
 using TimberApi.SceneSystem;
 
@@ -14,6 +15,7 @@ sealed class Configurator : IConfigurator {
   public void Configure(IContainerDefinition containerDefinition) {
     containerDefinition.Bind<DirectWaterServiceAccessor>().AsSingleton();
     containerDefinition.Bind<DirectSoilMoistureSystemAccessor>().AsSingleton();
+    containerDefinition.Bind<ThreadedLogsRecorder>().AsSingleton();
   }
 }